Add Settings to Enable/Disable Removal of Whitespace
Add Settings Window to Enable/Disable Removal of Whitespace.  In my
experience, there have been a few times while contributing to public
repos that I don’t want to be too disruptive to the files I modify by
removing the all the whitespace (tabs between statements).  The
addition of the setting option to enable/disable the removal of
whitespace will allow the user to disable the feature as needed.